User story 9 visit poi

+++ app/src/main/java/com/stevensadler/android/blocspot/api/model/database/table/Table.java
@@ -61,6 +61,6 @@ public abstract class Table {
     }
 
     protected static boolean getBoolean(Cursor cursor, String column) {
-        return getLong(cursor, column) == 1l;
+        return getInt(cursor, column) == 1;
     }
 }
